query-visualization/requirements.txt
oracledb
fastapi
uvicorn
pandas
plotly
python-dotenv
pytest
openai
